+++ transports/dns/src/lib.rs
@@ -60,15 +60,23 @@ use futures::{future::BoxFuture, prelude::*};
 use libp2p_core::{
     connection::Endpoint,
     multiaddr::{Multiaddr, Protocol},
-    transport::{ListenerEvent, TransportError},
+    transport::{ListenerId, TransportError, TransportEvent},
     Transport,
 };
 use parking_lot::Mutex;
 use smallvec::SmallVec;
 #[cfg(any(feature = "async-std", feature = "tokio"))]
 use std::io;
-use std::sync::Arc;
-use std::{convert::TryFrom, error, fmt, iter, net::IpAddr, str};
+use std::{
+    convert::TryFrom,
+    error, fmt, iter,
+    net::IpAddr,
+    ops::DerefMut,
+    pin::Pin,
+    str,
+    sync::Arc,
+    task::{Context, Poll},
+};
 #[cfg(any(feature = "async-std", feature = "tokio"))]
 use trust_dns_resolver::system_conf;
 use trust_dns_resolver::{proto::xfer::dns_handle::DnsHandle, AsyncResolver, ConnectionProvider};
@@ -174,7 +182,7 @@ where
 
 impl<T, C, P> Transport for GenDnsConfig<T, C, P>
 where
-    T: Transport + Send + 'static,
+    T: Transport + Send + Unpin + 'static,
     T::Error: Send,
     T::Dial: Send,
     C: DnsHandle<Error = ResolveError>,
@@ -182,38 +190,21 @@ where
 {
     type Output = T::Output;
     type Error = DnsErr<T::Error>;
-    type Listener = stream::MapErr<
-        stream::MapOk<
-            T::Listener,
-            fn(
-                ListenerEvent<T::ListenerUpgrade, T::Error>,
-            ) -> ListenerEvent<Self::ListenerUpgrade, Self::Error>,
-        >,
-        fn(T::Error) -> Self::Error,
-    >;
     type ListenerUpgrade = future::MapErr<T::ListenerUpgrade, fn(T::Error) -> Self::Error>;
     type Dial = future::Either<
         future::MapErr<T::Dial, fn(T::Error) -> Self::Error>,
         BoxFuture<'static, Result<Self::Output, Self::Error>>,
     >;
 
-    fn listen_on(
-        &mut self,
-        addr: Multiaddr,
-    ) -> Result<Self::Listener, TransportError<Self::Error>> {
-        let listener = self
-            .inner
+    fn listen_on(&mut self, addr: Multiaddr) -> Result<ListenerId, TransportError<Self::Error>> {
+        self.inner
             .lock()
             .listen_on(addr)
-            .map_err(|err| err.map(DnsErr::Transport))?;
-        let listener = listener
-            .map_ok::<_, fn(_) -> _>(|event| {
-                event
-                    .map(|upgr| upgr.map_err::<_, fn(_) -> _>(DnsErr::Transport))
-                    .map_err(DnsErr::Transport)
-            })
-            .map_err::<_, fn(_) -> _>(DnsErr::Transport);
-        Ok(listener)
+            .map_err(|e| e.map(DnsErr::Transport))
+    }
+
+    fn remove_listener(&mut self, id: ListenerId) -> bool {
+        self.inner.lock().remove_listener(id)
     }
 
     fn dial(&mut self, addr: Multiaddr) -> Result<Self::Dial, TransportError<Self::Error>> {
@@ -230,11 +221,23 @@ where
     fn address_translation(&self, server: &Multiaddr, observed: &Multiaddr) -> Option<Multiaddr> {
         self.inner.lock().address_translation(server, observed)
     }
+
+    fn poll(
+        self: Pin<&mut Self>,
+        cx: &mut Context<'_>,
+    ) -> Poll<TransportEvent<Self::ListenerUpgrade, Self::Error>> {
+        let mut inner = self.inner.lock();
+        Transport::poll(Pin::new(inner.deref_mut()), cx).map(|event| {
+            event
+                .map_upgrade(|upgr| upgr.map_err::<_, fn(_) -> _>(DnsErr::Transport))
+                .map_err(DnsErr::Transport)
+        })
+    }
 }
 
 impl<T, C, P> GenDnsConfig<T, C, P>
 where
-    T: Transport + Send + 'static,
+    T: Transport + Send + Unpin + 'static,
     T::Error: Send,
     T::Dial: Send,
     C: DnsHandle<Error = ResolveError>,
@@ -571,11 +574,10 @@ fn invalid_data(e: impl Into<Box<dyn std::error::Error + Send + Sync>>) -> io::E
 #[cfg(test)]
 mod tests {
     use super::*;
-    use futures::{future::BoxFuture, stream::BoxStream};
+    use futures::future::BoxFuture;
     use libp2p_core::{
         multiaddr::{Multiaddr, Protocol},
-        transport::ListenerEvent,
-        transport::TransportError,
+        transport::{TransportError, TransportEvent},
         PeerId, Transport,
     };
 
@@ -589,18 +591,18 @@ mod tests {
         impl Transport for CustomTransport {
             type Output = ();
             type Error = std::io::Error;
-            type Listener = BoxStream<
-                'static,
-                Result<ListenerEvent<Self::ListenerUpgrade, Self::Error>, Self::Error>,
-            >;
             type ListenerUpgrade = BoxFuture<'static, Result<Self::Output, Self::Error>>;
             type Dial = BoxFuture<'static, Result<Self::Output, Self::Error>>;
 
             fn listen_on(
                 &mut self,
                 _: Multiaddr,
-            ) -> Result<Self::Listener, TransportError<Self::Error>> {
+            ) -> Result<ListenerId, TransportError<Self::Error>> {
                 unreachable!()
+            }
+
+            fn remove_listener(&mut self, _: ListenerId) -> bool {
+                false
             }
 
             fn dial(&mut self, addr: Multiaddr) -> Result<Self::Dial, TransportError<Self::Error>> {
@@ -625,13 +627,20 @@ mod tests {
             fn address_translation(&self, _: &Multiaddr, _: &Multiaddr) -> Option<Multiaddr> {
                 None
             }
+
+            fn poll(
+                self: Pin<&mut Self>,
+                _: &mut Context<'_>,
+            ) -> Poll<TransportEvent<Self::ListenerUpgrade, Self::Error>> {
+                unreachable!()
+            }
         }
 
         async fn run<T, C, P>(mut transport: GenDnsConfig<T, C, P>)
         where
             C: DnsHandle<Error = ResolveError>,
             P: ConnectionProvider<Conn = C>,
-            T: Transport + Clone + Send + 'static,
+            T: Transport + Clone + Send + Unpin + 'static,
             T::Error: Send,
             T::Dial: Send,
         {