+++ muxers/mplex/CHANGELOG.md
@@ -1,4 +1,4 @@
-# 0.34.0 [unreleased]
+# 0.34.0
 
 - `Substream` now implements `AsyncRead` and `AsyncWrite`. See [PR 2706].
 - Update to `libp2p-core` `v0.34.0`