+++ protocols/gossipsub/CHANGELOG.md
@@ -1,8 +1,16 @@
-# 0.39.0 [unreleased]
+# 0.40.0 [unreleased]
+
+- Update to `libp2p-swarm` `v0.38.0`.
+
+# 0.39.0
 
 - Update to `libp2p-core` `v0.34.0`.
 
 - Update to `libp2p-swarm` `v0.37.0`.
+
+- Allow for custom protocol ID via `GossipsubConfigBuilder::protocol_id()`. See [PR 2718].
+
+[PR 2718]: https://github.com/libp2p/rust-libp2p/pull/2718/
 
 # 0.38.1