+++ protocols/identify/CHANGELOG.md
@@ -1,8 +1,14 @@
-# 0.37.0 [unreleased]
+# 0.38.0 [unreleased]
+
+- Update to `libp2p-swarm` `v0.38.0`.
+
+# 0.37.0
 
 - Update to `libp2p-core` `v0.34.0`.
 
 - Update to `libp2p-swarm` `v0.37.0`.
+
+- Extend log message on second identify push stream with peer ID.
 
 # 0.36.1