+++ swarm-derive/CHANGELOG.md
@@ -1,3 +1,9 @@
+# 0.28.0
+
+- Import `ListenerId` from `libp2p::core::transport`. See [PR 2652].
+
+[PR 2652]: https://github.com/libp2p/rust-libp2p/pull/2652
+
 # 0.27.2
 
 - Replace references of Protocol Handler with Connection Handler. See [PR 2640].